V1.3
The new version turns the code for getting the modifier for each person into a function, for ease of reading. It also checks for validity whenever it asks for user input, then prints an appropriate error message depending on what was wrong with the input. Finally, the new version also prints the items that are on the table whenever there is a change to them.