V1.5
The new version combines the "Add" and "Remove" actions into one action; "Edit." Edit first checks if the item you entered exists, then if so, toggles it. The new version also includes assertions, which replace the validity checks in V1.4. The code for the get_modifier function has also been turned into less code and a dictionary, which can be found at the top of the program