V1.6
V1.6 adds the ability to turn the table manually using the "Turn" command. Many changes were made to formatting to comply with PEP 8. The "Edit" command was renamed "Toggle." You can now change the position of an item during the run by using the new "Edit" command. When you enable an item, the program will automatically run the edit command.

Co-Authored-By: Graham Lea <graham@grahamlea.com>